[WIN] Solution has been found!
In fact, it wasn't the way I implemented the Alarms which
was causing this bug but the device on which I tested.

I tried all implementations on a Samsung device (my test device)
and a Nexus device. Five days later, nothing was running anymore
on the Samsung. However, on the Nexus everything was fine,
all my 4 apps were still triggering the alarms

Samsung has a strong battery optimization policy.
When an app has not been in foreground for a while,
Samsung kills all its background task.
To avoid this behavior, the user has to:

* Go to "device maintenance"
* click on "Battery"
* Scroll down to "Unmonitored apps"
* Click on the up right button, and click on "add"
* Add your app.